35.2/requirements.txt
absl-py==2.3.1
aenum==3.1.16
affine==2.4.0
aiosqlite==0.22.1
alembic==1.18.1
altair==6.0.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.1
asttokens==3.0.1
async-generator==1.10
asyncpg==0.31.0
attrs==25.4.0
backports.tarfile==1.2.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.47
botocore==1.42.47
brotli==1.2.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.2
celery==5.6.2
celery-redis-sentinel==0.3.0
certifi==2026.1.4
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.4
crashtest==0.4.1
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.3.1
distlib==0.4.0
distro==1.9.0
Django==5.2.11
django-celery-beat==2.8.1
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
dnspython==2.8.0
dulwich==0.24.10
earthengine-api==1.7.13
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.1
flake8==7.3.0
fonttools==4.61.0
fsspec==2026.1.0
GeoAlchemy2==0.14.2
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.29.0
google-api-python-client==2.190.0
googleapis-common-protos==1.72.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-cloud-translate==3.24.0
google-cloud-vision==3.12.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==22.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
immutabledict==4.2.2
importlib_metadata==8.7.0
iniconfig==2.3.0
installer==0.7.0
ipython_pygments_lexers==1.1.1
ipython==9.10.0
isort==8.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.2
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
langchain-core==1.1.3
langchain-openai==1.1.1
langsmith==0.4.59
lazy_loader==0.4
librt==0.8.1
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
mypy==1.19.1
narwhals==2.13.0
netCDF4==1.7.4
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
openai==2.9.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.2
orjson==3.11.5
ortools==9.14.6206
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandas-stubs==2.2.3.241126
parso==0.8.6
passlib==1.7.4
pathspec==1.0.4
patsy==1.0.2
pbs-installer==2025.12.5
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
poetry-plugin-export==1.9.0
prompt_toolkit==3.0.52
protobuf==6.31.1
proto-plus==1.27.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1_modules==0.4.2
pyasn1==0.6.1
pyclipper==1.4.0
pycodestyle==2.14.0
pycparser==2.23
pydantic_core==2.41.5
pydantic==2.12.5
pydantic-settings==2.1.0
pydeck==0.9.1
pydyf==0.12.1
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.2.5
pypdf==6.6.0
PyPDF2==3.0.1
pypdfium2==5.3.0
pyphen==0.17.2
pyproj==3.7.2
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==7.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-pptx==1.0.2
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
rasterio==1.4.4
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
reportlab==4.0.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rioxarray==0.19.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.40.0
sentinelhub==3.11.4
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.6
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.20
tiktoken==0.12.0
tinycss2==1.5.1
toml==0.10.2
tomli_w==1.2.0
tomli==2.4.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
trove-classifiers==2025.12.1.14
types-certifi==2021.10.8.3
types-requests==2.32.4.20260107
types-urllib3==1.26.25.14
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
utm==0.8.1
uuid_utils==0.12.0
uvicorn==0.24.0
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
weasyprint==60.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
wsproto==1.3.2
xarray==2026.1.0
xlsxwriter==3.2.9
zipp==3.23.0
zopfli==0.4.0
zstandard==0.25.0
